Changed boundary function return type
Allow getBoundaryTypeFromGeocoderAddressTypes() to return undefined if the geocode result type does't match any of the available mappings. The fact that it doesn't match most likely indicates that there are no boundaries that would be returned for the geocoded location, which would be good information to have when trying to determine if we're searching for a type of location that actually would have bounds. If we were searching for an address we wouldn't really want this to implicitly return "neighborhood", as it previously did. If in the future there is some scenario where it would make sense to use a default value it would probably be better to have the caller explicitly use that if nothing is returned instead.