Got basic address/no-boundary search working
If the type returned by the geocoder service is not a type of boundary that we support, we're first assuming that the user requested an address, and trying to find a listing that matches the place_id for that address and returning it in the response. If we don't find a listing then we just return the result from the geocoder and allow the front end to adjust the map then do a search on the viewport bounds instead. This also works well if we simply don't have a boundary for the place that was searched, so we're removing the code that was returning a 404 when no boundary was found in favor of this solution.