Changed interfaces to types
We were getting an error about extending Context incorrectly after switching Typescript to strict mode. We want to switch to using types instead of interfaces anyway so decided to just do that rather than resolve the issue with extending the interface in the exact right way by using Omit<> on the offending query attribute.