Bug fixes and refactoring for buildFilterQueries
Changes to add Zod validation made it so that the params coming into buildFilterQueries are now their correct types, e.g., Number, Boolean, etc. This introduced a bug with the boolean params since they aren't strings anymore. We have to check the actual type for boolean params because if the value is false it means to exclude records. A simple truthiness check wouldn't accomplish that.