Replaced mongodb-memory-server
I was having lots of difficulty getting mongodb-memory-server to work well with Docker. It has a dependency on the MongoDB binary that it tries to download automatically depending on your Docker image distro and the architecture of the host machine. It was really painful trying to get the combination of those things right. I decided to ditch it in favor of just using an official MongoDB image with tmpfs. This way it still runs in memory so we can keep tests fast, plus it's the real MongoDB, so will be closer to production. The setup for Jest is much simpler than it was for mongodb-memory-server too.