Removed build command from Dockerfile
It seems that Jest uses the build files in the dist/ directory if we run it inside the container. This was causing test changes to never update, which was very confusing. Removing the build command seems to have helped the issue. We will want to do a proper production build for the container later on down the road and make sure each process is separate.