Narrowing scope of test files
We were having problems with tests not updating earlier. It was because Jest was running the compiled versions of the tests that were created when the image ran a build. We've removed the build but I want to make sure that nothing will be ambiguous like this again. I've moved all tests inside the tests directory so they won't be part of the build again. I've also added config to make sure Jest only runs Typescript files that are in the tests directory.