Refactored app to use repository pattern
I refactored the app data layer to use a repository pattern instead of calling the mongoose models directly. The repository will be added to the global context object. The main reason for this change was that I wanted to be able to test parts of the app in isolation without requiring a real database connection. This way I can switch out the repositories context with a mock that conforms to the same interface. This will also make it easier to move to a different database in the future if desirable.

I also renamed listing_search_helpers to listingSearchService, mainly because that name seems more appropriate for what that module actually does.