Added script to seed mongo container
Added a script to seed the dev and test mongodb containers with data when they're first started. This way there doesn't need to be an extra step involved for setting up the project. It also makes the tests simpler because we can just use the data that is in the test container rather than having to re-create it inside the tests each time they run.