Refactored geocode endpoint
I didn't like the way I had separated out the concerns between the controller and the service. I want the controller to be responsible for calling the correct view function, not the service. I also realized that the logic I had put together was way too complicated than it needed to be, so I simplified things. Also added some doc strings to the service functions to make their purpose more clear.