Added app-test container
I wanted the app container used for testing to know that it depends on mongodb-test when running the test script so that the person running it doesn't have to remember to make sure it's already running. Seems that you can't do that on the command line so I just made a dedicated service for the test app.