Added createListing function
We were running into issues where MongoDB was returning a duplicate key error for the listing slug field. This was because Jest executes test files in parallel by default, which caused our slug generation hook to run into a race condition when attempting to generate the slug. The hook attempts to check for existing slugs before saving but apparently this was not enough when there are multiple saves happening in parallel. The solution I chose to resolve this was to retry the save in the createListing function. It's not perfect is simple and will likely be good enough for most cases. We don't actually create or update listings within the service itself, so a more robust solution doesn't feel necessary yet.