Making filters test it's own file
We originally implemented the filter tests as a function that can be re-used for each endpoint that has filters. That was nice but kind of feels like overkill. For now we're just going to test filters with the bounds endpoint and assume that if they pass that they would also pass for the geocode endpoint. The two endpoints use the exact same code for the queries so it seems like a reasonable assumption to make. We can go back to the other way later if that assumption doesn't pan out.