Upgraded Mongoose to latest version
There were Typescript errors for the ListingResultProjectionFields and ListingDetailResultProjectionFields objects. The type appears to be more strict when passing the object to find, findOne, findById, etc., which requires ProjectionType<IListing>, but it seems that ProjectionType<IListing> is to strict for the $project part of an aggregate query. The best solution I could find was to use satisfies ProjectionType<IListing>. That seems to be sufficient for both use cases.