Boundary outside of bounds returns listings
Bug fix and new tests for an issue where the dragging the viewport so that the boundary is no longer visible was still returning listings. The expected behavior was that no listings would be returned since the boundary was turned on, and it normally removes listings when part of the boundary is outside the viewport. The main issue was that the getBoundaryGeometryWithBounds() function would return boundary.geometry if the removePartsOfBoundaryOutsideOfBounds() returned undefined. It should have just returned the value that intersect returns, which is null. The controller/view should expect no boundary to be returned if it receives null from this function. The null values means there is no intersection between the two boundaries, which to us means that the boundary is completely outside the map viewport.