Narrowing the type for place_id lookup
I would like to be able to create a mock context for testing these functions but I don't want to have to create an entire Context object since it has a lot of different properties that aren't necessary for testing them. So, I'm trying out the idea of creating specific types that only include the interface that is necessary for the functions to do what they need to do.