Changed /listing endpoint param
We're looking up the listing by slug rather than ID now because that's what the client gets in the url. Normally we might want to put this in a new version (/v2/listing) of the routes but because there's nothing besides our one app that consumes the endpoint it doesn't really matter.